+++ src/views/Account/index.js
@@ -10,7 +10,6 @@ import Button from '../../components/Button';
 import CopyToClipboard from '../../components/CopyToClipboard';
 import AccountBalances from './AccountBalances';
 import AccountTransactions from './AccountTransactions';
-import AccountInteractions from './AccountInteractions';
 import arrowUp from '../../assets/arrow-up.svg';
 import exchangeIcon from '../../assets/exchange-icon.svg';
 import qrCode from '../../assets/qr-code-transparent.svg';
@@ -147,11 +146,6 @@ class Account extends Component {
                   exact
                   path={`${this.props.match.url}/transactions`}
                   component={AccountTransactions}
-                />
-                <Route
-                  exact
-                  path={`${this.props.match.url}/interactions`}
-                  component={AccountInteractions}
                 />
                 <Route render={() => <Redirect to={this.props.match.url} />} />
               </Switch>